Fix matcher behavior, restructure code
