+++ moto/secretsmanager/list_secrets/filters.py
@@ -1,20 +1,20 @@
-def matcher(pattern, str):
+def _matcher(pattern, str):
     for word in pattern.split(" "):
-        if word in str:
-            return True
-    return False
+        if word not in str:
+            return False
+    return True
 
 
 def name(secret, names):
     for n in names:
-        if matcher(n, secret["name"]):
+        if _matcher(n, secret["name"]):
             return True
     return False
 
 
 def description(secret, descriptions):
     for d in descriptions:
-        if matcher(d, secret["description"]):
+        if _matcher(d, secret["description"]):
             return True
     return False
 
@@ -22,7 +22,7 @@ def description(secret, descriptions):
 def tag_key(secret, tag_keys):
     for k in tag_keys:
         for tag in secret["tags"]:
-            if matcher(k, tag["Key"]):
+            if _matcher(k, tag["Key"]):
                 return True
     return False
 
@@ -30,7 +30,7 @@ def tag_key(secret, tag_keys):
 def tag_value(secret, tag_values):
     for v in tag_values:
         for tag in secret["tags"]:
-            if matcher(v, tag["Value"]):
+            if _matcher(v, tag["Value"]):
                 return True
     return False
