+++ tests/test_secretsmanager/test_secretsmanager.py
@@ -15,10 +15,6 @@ from six import b
 DEFAULT_SECRET_NAME = "test-secret"
 
 
-def boto_client():
-    return boto3.client("secretsmanager", region_name="us-west-2")
-
-
 @mock_secretsmanager
 def test_get_secret_value():
     conn = boto3.client("secretsmanager", region_name="us-west-2")
@@ -461,214 +457,6 @@ def test_describe_secret_that_does_not_match():
 
     with assert_raises(ClientError):
         result = conn.get_secret_value(SecretId="i-dont-match")
-
-
-@mock_secretsmanager
-def test_list_secrets_empty():
-    conn = boto_client()
-
-    secrets = conn.list_secrets()
-
-    assert secrets["SecretList"] == []
-
-
-@mock_secretsmanager
-def test_list_secrets():
-    conn = boto_client()
-
-    conn.create_secret(Name="test-secret", SecretString="foosecret")
-
-    conn.create_secret(
-        Name="test-secret-2",
-        SecretString="barsecret",
-        Tags=[{"Key": "a", "Value": "1"}],
-    )
-
-    secrets = conn.list_secrets()
-
-    assert secrets["SecretList"][0]["ARN"] is not None
-    assert secrets["SecretList"][0]["Name"] == "test-secret"
-    assert secrets["SecretList"][1]["ARN"] is not None
-    assert secrets["SecretList"][1]["Name"] == "test-secret-2"
-    assert secrets["SecretList"][1]["Tags"] == [{"Key": "a", "Value": "1"}]
-
-
-@mock_secretsmanager
-def test_list_secrets_with_name_filter():
-    conn = boto_client()
-
-    conn.create_secret(Name="foo", SecretString="secret")
-    conn.create_secret(Name="bar", SecretString="secret")
-
-    secrets = conn.list_secrets(Filters=[
-        {
-            "Key": "name",
-            "Values": ["foo"]
-        }
-    ])
-
-    secret_names = list(map(lambda s: s["Name"], secrets["SecretList"]))
-    assert secret_names == ["foo"]
-
-
-@mock_secretsmanager
-def test_list_secrets_with_tag_key_filter():
-    conn = boto_client()
-
-    conn.create_secret(Name="foo", SecretString="secret", Tags=[{"Key": "baz", "Value": "1"}])
-    conn.create_secret(Name="bar", SecretString="secret")
-
-    secrets = conn.list_secrets(Filters=[
-        {
-            "Key": "tag-key",
-            "Values": ["baz"]
-        }
-    ])
-
-    secret_names = list(map(lambda s: s["Name"], secrets["SecretList"]))
-    assert secret_names == ["foo"]
-
-
-@mock_secretsmanager
-def test_list_secrets_with_tag_value_filter():
-    conn = boto_client()
-
-    conn.create_secret(Name="foo", SecretString="secret", Tags=[{"Key": "1", "Value": "baz"}])
-    conn.create_secret(Name="bar", SecretString="secret")
-
-    secrets = conn.list_secrets(Filters=[
-        {
-            "Key": "tag-value",
-            "Values": ["baz"]
-        }
-    ])
-
-    secret_names = list(map(lambda s: s["Name"], secrets["SecretList"]))
-    assert secret_names == ["foo"]
-
-
-@mock_secretsmanager
-def test_list_secrets_with_description_filter():
-    conn = boto_client()
-
-    conn.create_secret(Name="foo", SecretString="secret", Description="baz qux")
-    conn.create_secret(Name="bar", SecretString="secret")
-
-    secrets = conn.list_secrets(Filters=[
-        {
-            "Key": "description",
-            "Values": ["baz"]
-        }
-    ])
-
-    secret_names = list(map(lambda s: s["Name"], secrets["SecretList"]))
-    assert secret_names == ["foo"]
-
-
-@mock_secretsmanager
-def test_list_secrets_with_all_filter():
-    conn = boto_client()
-
-    conn.create_secret(Name="foo", SecretString="secret")
-    conn.create_secret(Name="bar", SecretString="secret", Description="foo")
-    conn.create_secret(Name="baz", SecretString="secret", Tags=[{"Key": "foo", "Value": "1"}])
-    conn.create_secret(Name="qux", SecretString="secret", Tags=[{"Key": "1", "Value": "foo"}])
-    conn.create_secret(Name="multi", SecretString="secret", Tags=[{"Key": "foo", "Value": "foo"}])
-    conn.create_secret(Name="none", SecretString="secret")
-
-    secrets = conn.list_secrets(Filters=[
-        {
-            "Key": "all",
-            "Values": ["foo"]
-        }
-    ])
-
-    secret_names = list(map(lambda s: s["Name"], secrets["SecretList"]))
-    assert secret_names == ["foo", "bar", "baz", "qux", "multi"]
-
-
-@mock_secretsmanager
-def test_list_secrets_with_invalid_filter():
-    conn = boto_client()
-
-    with assert_raises(ClientError) as ire:
-        conn.list_secrets(Filters=[
-            {
-                "Key": "invalid",
-                "Values": ["foo"]
-            }
-        ])
-
-    ire.exception.response["Error"]["Code"].should.equal("InvalidRequestException")
-    # TODO find out what error is raised
-    ire.exception.response["Error"]["Message"].should.equal(
-        "???."
-    )
-
-
-@mock_secretsmanager
-def test_list_secrets_with_multiple_filters():
-    conn = boto_client()
-
-    # Multiple filters combine with an implicit AND operator
-    conn.create_secret(Name="foo", SecretString="secret", Tags=[{"Key": "right", "Value": "right"}])
-    conn.create_secret(Name="bar", SecretString="secret", Tags=[{"Key": "right", "Value": "wrong"}])
-    conn.create_secret(Name="baz", SecretString="secret", Tags=[{"Key": "wrong", "Value": "right"}])
-    conn.create_secret(Name="qux", SecretString="secret", Tags=[{"Key": "wrong", "Value": "wrong"}])
-
-    secrets = conn.list_secrets(Filters=[
-        {
-            "Key": "tag-key",
-            "Values": ["right"]
-        },
-        {
-            "Key": "tag-value",
-            "Values": ["right"]
-        }
-    ])
-
-    secret_names = list(map(lambda s: s["Name"], secrets["SecretList"]))
-    assert secret_names == ["foo"]
-
-
-@mock_secretsmanager
-def test_list_secrets_with_filter_with_multiple_values():
-    conn = boto_client()
-
-    conn.create_secret(Name="foo", SecretString="secret")
-    conn.create_secret(Name="bar", SecretString="secret")
-    conn.create_secret(Name="baz", SecretString="secret")
-
-    secrets = conn.list_secrets(Filters=[
-        {
-            "Key": "name",
-            "Values": ["foo", "bar"]
-        }
-    ])
-
-    secret_names = list(map(lambda s: s["Name"], secrets["SecretList"]))
-    assert secret_names == ["foo", "bar"]
-
-
-@mock_secretsmanager
-def test_list_secrets_with_filter_with_value_with_multiple_words():
-    conn = boto_client()
-
-    conn.create_secret(Name="foo", SecretString="secret", Description="one two")
-    conn.create_secret(Name="bar", SecretString="secret", Description="one and two")
-    conn.create_secret(Name="baz", SecretString="secret", Description="one")
-    conn.create_secret(Name="qux", SecretString="secret", Description="two")
-    conn.create_secret(Name="none", SecretString="secret", Description="unrelated")
-
-    secrets = conn.list_secrets(Filters=[
-        {
-            "Key": "description",
-            "Values": ["one two"]
-        }
-    ])
-
-    secret_names = list(map(lambda s: s["Name"], secrets["SecretList"]))
-    assert secret_names == ["foo", "bar", "baz", "qux"]
 
 
 @mock_secretsmanager
